+++ src/main/kotlin/com/ChickenKitchen/app/serviceImpl/menu/NutrientServiceImpl.kt
@@ -1,5 +1,8 @@
 package com.ChickenKitchen.app.serviceImpl.menu
 
+import com.ChickenKitchen.app.handler.NutrientHasMenuItemsException
+import com.ChickenKitchen.app.handler.NutrientNameExistException
+import com.ChickenKitchen.app.handler.NutrientNotFoundException
 import com.ChickenKitchen.app.model.dto.request.CreateNutrientRequest
 import com.ChickenKitchen.app.model.dto.request.UpdateNutrientRequest
 import com.ChickenKitchen.app.model.dto.response.NutrientDetailResponse
@@ -24,11 +27,16 @@ class NutrientServiceImpl(
     }
 
     override fun getById(id: Long): NutrientDetailResponse {
-        val entity = nutrientRepository.findById(id).orElseThrow { NoSuchElementException("Nutrient with id $id not found") }
+        val entity = nutrientRepository.findById(id)
+            .orElseThrow { NutrientNotFoundException("Nutrient with id $id not found") }
         return entity.toNutrientDetailResponse()
     }
 
     override fun create(req: CreateNutrientRequest): NutrientDetailResponse {
+        val existedByName = nutrientRepository.findByName(req.name)
+        if(existedByName != null){
+            throw NutrientNameExistException("Nutrient Name is already exists")
+        }
         val entity = Nutrient(
             name = req.name,
             baseUnit = req.baseUnit,
@@ -38,7 +46,8 @@ class NutrientServiceImpl(
     }
 
     override fun update(id: Long, req: UpdateNutrientRequest): NutrientDetailResponse {
-        val current = nutrientRepository.findById(id).orElseThrow { NoSuchElementException("Nutrient with id $id not found") }
+        val current = nutrientRepository.findById(id)
+            .orElseThrow { NutrientNotFoundException("Nutrient with id $id not found") }
         val updated = Nutrient(
             id = current.id,
             name = req.name ?: current.name,
@@ -50,7 +59,14 @@ class NutrientServiceImpl(
     }
 
     override fun delete(id: Long) {
-        val current = nutrientRepository.findById(id).orElseThrow { NoSuchElementException("Nutrient with id $id not found") }
+        val current = nutrientRepository.findById(id)
+            .orElseThrow { NutrientNotFoundException("Nutrient with id $id not found") }
+
+
+        if (current.menuItemNutrients.isNotEmpty()) {
+            throw NutrientHasMenuItemsException("Cannot delete Nutrient with id $id: it is used in ${current.menuItemNutrients.size} menu items")
+        }
+
         nutrientRepository.delete(current)
     }
 }
